requirements.txt
numpy
pandas
matplotlib
seaborn
Pillow
scikit-learn
scipy
PyYAML
torch
torchvision
